+++ problem/subset_sum.go
@@ -1,7 +1,10 @@
 package problem
 
 import (
+	"slices"
+
 	"github.com/roidaradal/fn/list"
+	"github.com/roidaradal/opt/data"
 	"github.com/roidaradal/opt/discrete"
 	"github.com/roidaradal/opt/fn"
 )
@@ -12,6 +15,10 @@ func NewSubsetSum(variant string, n int) *discrete.Problem {
 	switch variant {
 	case "basic":
 		return subsetSum(name)
+	case "max_sum":
+		return maxSumMultipleSubsetSum(name)
+	case "max_min":
+		return maxMinMultipleSubsetSum(name)
 	default:
 		return nil
 	}
@@ -45,5 +52,73 @@ func subsetSum(name string) *discrete.Problem {
 		}
 		return discrete.Score(cfg.Target - total)
 	}
+	return p
+}
+
+// Common steps for creating multiple subset sum problem
+func newMultipleSubsetSumProblem(name string) (*discrete.Problem, *data.Numbers) {
+	cfg := data.NewNumbers(name)
+	if cfg == nil || cfg.Target == 0 || cfg.NumBins == 0 {
+		return nil, nil
+	}
+
+	p := discrete.NewProblem(name)
+	p.Type = discrete.Partition
+	p.Goal = discrete.Maximize
+
+	p.Variables = discrete.Variables(cfg.Numbers)
+	p.AddVariableDomains(discrete.RangeDomain(0, cfg.NumBins)) // 0 = not in bin
+
+	validDomain := discrete.RangeDomain(1, cfg.NumBins)
+	p.AddUniversalConstraint(func(solution *discrete.Solution) bool {
+		// Total of numbers in each bin must not exceed capacity
+		for _, partition := range fn.AsPartition(solution, validDomain) {
+			total := list.Sum(list.MapList(partition, cfg.Numbers))
+			if total > cfg.Target {
+				return false
+			}
+		}
+		return true
+	})
+
+	p.SolutionStringFn = fn.StringPartition(validDomain, cfg.Numbers)
+	return p, cfg
+}
+
+// Max-Sum Multiple Subset Sum
+func maxSumMultipleSubsetSum(name string) *discrete.Problem {
+	p, cfg := newMultipleSubsetSumProblem(name)
+	if p == nil || cfg == nil {
+		return nil
+	}
+
+	validDomain := discrete.RangeDomain(1, cfg.NumBins)
+	p.ObjectiveFn = func(solution *discrete.Solution) discrete.Score {
+		// Sum up total numbers across bins
+		total := 0
+		for _, partition := range fn.AsPartition(solution, validDomain) {
+			total += list.Sum(list.MapList(partition, cfg.Numbers))
+		}
+		return discrete.Score(total)
+	}
+	return p
+}
+
+// Max-Min Multiple Subset Sum
+func maxMinMultipleSubsetSum(name string) *discrete.Problem {
+	p, cfg := newMultipleSubsetSumProblem(name)
+	if p == nil || cfg == nil {
+		return nil
+	}
+
+	validDomain := discrete.RangeDomain(1, cfg.NumBins)
+	p.ObjectiveFn = func(solution *discrete.Solution) discrete.Score {
+		// Find min total numbers across bins
+		totals := list.Map(fn.AsPartition(solution, validDomain), func(partition []discrete.Variable) int {
+			return list.Sum(list.MapList(partition, cfg.Numbers))
+		})
+		return discrete.Score(slices.Min(totals))
+	}
+
 	return p
 }